Added Helper.h include file
