+++ src/main/include/helper.h
@@ -1,0 +1,84 @@
+#ifndef HELPERS_H
+#include <cmath>
+#include "constants.h"
+
+
+// magnitude of vector to convert from Cartesian to Polar
+double magnitude(double x, double y) 
+{
+    return sqrt(x * x + y * y);
+}
+
+// depending on the angles, the optimal turn may reverse motor direction
+// Returns 1 or -1 based on optimal turn angle
+// Takes initial and desired angle
+int magnitudeOptimization(double initialAngle, double finalAngle)
+{
+    double posDiff = fabs(fmod(finalAngle - initialAngle, 360.0));
+    if (posDiff > 90.0 && posDiff < 270.0)
+    {
+        return -1;
+    }
+    return 1;
+}
+
+// ideally, the wheel will only ever turn 90 degrees clockwise or counterclockwise and change drive motor direction to reach desired angle
+// Returns optimal turn angle based 
+// Takes initial and desired angle
+double angleOptimisation(double initialAngle, double finalAngle)
+{
+    double diff = fmod(finalAngle - initialAngle, 360.0);
+    double posDiff = fabs(diff);
+    if (posDiff > 90.0 && posDiff < 270.0)
+    {
+        diff = fmod(diff + 180.0, 360.0);
+    }
+    if (fabs(diff) <= 90.0)
+    {
+        return diff;
+    }
+    else if (diff >= 270.0)
+    {
+        return diff - 360.0;
+    }
+    return diff + 360.0;
+}
+
+// Given an array of values, find the largest member greater than 1
+// if all members are less than one, return 1
+// else, return largest value
+double findMax(double arr[], int len) // finds max value in array
+{
+    double Max = 1.0;
+    for (int i = 0; i < len; i++)
+    {
+        Max = fmax(Max, arr[i]);
+    }
+    return Max;
+}
+
+// Convert radian to degree
+double getDegree(double x)
+{
+    return x * 180.0 / M_PI;
+}
+
+// Convert degree to radian
+double getRadian(double x)
+{
+    return x * M_PI / 180.0;
+}
+
+// Deadband for joystick
+// prevents drift at values close to 0
+double deadband(double joystickInput)
+{
+    if (abs(joystickInput) <= mathConst::deadband)
+    {
+        return 0.0;
+    }
+    return joystickInput;
+}
+
+
+#endif
